Some of the models used.
The ring gear is a scaled version of MRBADDELEY's R2 gear. The Motor gear is scaled as well but keeps the correct motor mount size.

The amazon motor bracket is modified specifically for the amazon motor.

The dome adapter is a variation on the common dome adapter for the 14 inch lazy susan but it includes a platform for mounting an arduino and a 1000mAh 3 cell lipo for powering it.

The leg lock is for locking chopper in 3 leg or standing mode.